Add files via upload
The goal of this assignment is to understand and practice using a high-level language (Python) to interact with a database via ODBC.
Your goal is to write a program that runs from the command line. It interacts with an end-user to insert, delete, and update records as well as handle queries on the data.